backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4

# Machine Learning
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
camelot-py[cv]==0.11.0
pytesseract==0.3.10
opencv-python-headless==4.10.0.82
Pillow==10.4.0
python-docx==1.1.2

# Data Storage & Caching
redis==5.0.7
aioredis==2.0.1

# HTTP and API utilities
httpx==0.27.0
requests==2.32.3

# Security and validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Configuration and Environment
python-dotenv==1.0.1
PyYAML==6.0.1

# Logging and Monitoring
structlog==24.2.0

# Geographic services
geopy==2.4.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Development
black==24.4.2
isort==5.13.2
flake8==7.1.0